Move string and table wrappers into own files
